### 2025-02-20 #7815 @editorialbot
[PRE REVIEW]: omock: A R package for Mock Data Generation for the Observational Medical Outcomes Partnership common data model
**Submitting author:** @ilovemane (<a href="http://orcid.org/0000-0002-9517-8834">Mike Du</a>)
**Repository:** https://github.com/OHDSI/omock
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.2
