### 2025-04-04 faaa9f2
Update checks.R
