### 2025-04-04 #7815 @ilovemane
Hi, @jromanowska

Thanks for your time to be the editor of my paper. I made a mistake to the title. "mock: A R ..." it should be "omock: A R.." as my package is called omock. Can you edit this for me please?

thanks,

Mike
