### 2025-04-07 #7815 @jromanowska
@ilovemane - could you fix the missing DOIs, as [mentioned above](https://github.com/openjournals/joss-reviews/issues/7815#issuecomment-2670904658), while I'm finding reviewers?
