### 2025-04-08 #7815 @ilovemane
> [@ilovemane](https://github.com/ilovemane) - could you fix the missing DOIs, as [mentioned above](https://github.com/openjournals/joss-reviews/issues/7815#issuecomment-2670904658), while I'm finding reviewers?

I have filled in the missing doi now and merged it to main. @jromanowska
