### 2025-04-11 #7815 @chitrams
> 👋 [@chitrams](https://github.com/chitrams), [@louisahsmith](https://github.com/louisahsmith), & [@nrennie](https://github.com/nrennie), would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I'm happy to review, unfortunately I'll only have capacity from 1 May. Does this work for everyone? Thank you :)
