+++ R/mockObservationPeriod.R
@@ -55,10 +55,12 @@ mockObservationPeriod <- function(cdm,
 
 
     # generate observation date from dob
+    tictoc::tic()
     observationDate <- obsDate(dob = dob, max = max(
       as.Date("2020-01-01"),
       max(as.Date(dob))
     ))
+    tictoc::toc()
 
     person_id <- cdm$person |>
       dplyr::select(person_id) |>
@@ -94,19 +96,11 @@ mockObservationPeriod <- function(cdm,
 
 # function to generate mock observational period date from a list of dob
 obsDate <- function(dob = dob, max = "2020-01-01") {
-  # Initialise vector for output
-  start <- rep(as.Date(NA), length(dob))
-  end <- rep(as.Date(NA), length(dob))
-  # generate obs start and end date
-  for (i in seq_along(dob)) {
-    start[i] <- sample(seq(as.Date(dob[i]), as.Date(max),
-      by =
-        "day"
-    ), 1)
-    end[i] <- sample(seq(as.Date(start[i]), as.Date(max),
-      by =
-        "day"
-    ), 1)
-  }
+  #
+  r1 <- runif(n = length(dob))
+  start <- dob + round((as.Date(max) - dob) * r1)
+  r2 <- runif(n = length(dob))
+  end <- start + round((as.Date(max) - start) * r2)
+
   list(start, end)
 }
