### 2025-05-02 #7815 @chitrams
Hi @jromanowska, my work which I expected to die down from 1 May, is still pretty hectic. A few things have come up for our group. I estimate I'd need another two weeks before I have time to review this package. Sorry for the delay :( If it's too far in the future and you want to find another reviewer sooner rather than later, please go ahead. I'd love to review omock when my work slows down though! Thank you and sorry again for the inconvenience.
