### 2025-05-08 #8178 @editorialbot
[REVIEW]: mock: A R package for Mock Data Generation for the Observational Medical Outcomes Partnership common data model
**Submitting author:** @ilovemane (<a href="http://orcid.org/0000-0002-9517-8834">Mike Du</a>)
**Repository:** https://github.com/OHDSI/omock
**Branch with paper.md** (empty if default branch): main
**Version:** v0.5.0.1
