### 2025-05-15 #8178 @ilovemane
@jromanowska Hi, I am not sure hows this works. When would I need to start addressing the comment? I can see some boxes are not ticked above. Will I get a notification when the reviewers finish the reviews?
