### 2025-05-16 #8178 @jromanowska
> [@jromanowska](https://github.com/jromanowska) Hi, I am not sure hows this works. When would I need to start addressing the comment? I can see some boxes are not ticked above. Will I get a notification when the reviewers finish the reviews?

@ilovemane, the reviewers have started to look at the package, as you can clearly see from the boxes that are being ticked off. This thread is for communication between you and the reviewers. You can read [the guidelines for reviewers](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) if you're in doubt, but I'm sure they'll let you know if they encounter any problems.
