### 2025-06-04 e53c71f
Update mockObservationPeriod.R
