### 2025-06-04 cc87d36
Update mockObservationPeriod.R
