+++ R/mockObservationPeriod.R
@@ -25,13 +25,15 @@
 #'
 #' # View the generated observation period data
 #' print(cdm$observation_period)
-mockObservationPeriod <- function(cdm,
-                                  seed = NULL) {
+mockObservationPeriod <- function(cdm, seed = NULL) {
   checkInput(cdm = cdm)
-  if (nrow(cdm$observation_period) == 0 &
-    nrow(cdm$person) != 0) {
+
+  if (nrow(cdm$person) != 0) {
     if (!is.null(seed)) {
       set.seed(seed = seed)
+    }
+    if (nrow(cdm$observation_period) != 0) {
+      cli::cli_warn("The observation period table have been overwrite.")
     }
     # pull date of birth from person table
     dob <- cdm$person |>
@@ -56,10 +58,7 @@ mockObservationPeriod <- function(cdm,
 
     # generate observation date from dob
 
-    observationDate <- obsDate(dob = dob, max = max(
-      as.Date("2020-01-01"),
-      max(as.Date(dob))
-    ))
+    observationDate <- obsDate(dob = dob, max = max(as.Date("2020-01-01"), max(as.Date(dob))))
 
 
     person_id <- cdm$person |>
@@ -81,14 +80,16 @@ mockObservationPeriod <- function(cdm,
 
 
     cdm <-
-      omopgenerics::insertTable(
-        cdm = cdm,
-        name = "observation_period",
-        table = observationPeriod
-      )
-  }
+      omopgenerics::insertTable(cdm = cdm,
+                                name = "observation_period",
+                                table = observationPeriod)
 
+
+
+  }
   return(cdm)
+
+
 }
 
 
