### 2025-06-04 483751b
Revert "Update mockObservationPeriod.R"
This reverts commit cc87d36608ec7210604c165a2f3ea6cb94f617c4.
