### 2025-06-05 fef471a
Update mockObservationPeriod.R
