+++ R/mockObservationPeriod.R
@@ -28,8 +28,12 @@
 mockObservationPeriod <- function(cdm,
                                   seed = NULL) {
   checkInput(cdm = cdm)
-  if (nrow(cdm$observation_period) == 0 &
-    nrow(cdm$person) != 0) {
+  if (nrow(cdm$person) != 0) {
+
+   if(nrow(cdm$observation_period) == 0) {
+     cli::cli_inform("The observation period table has been overwritten.")
+   }
+
     if (!is.null(seed)) {
       set.seed(seed = seed)
     }
