### 2025-06-05 c1f1dd9
Update mockObservationPeriod.R
