+++ R/mockObservationPeriod.R
@@ -30,7 +30,7 @@ mockObservationPeriod <- function(cdm,
   checkInput(cdm = cdm)
   if (nrow(cdm$person) != 0) {
 
-   if(nrow(cdm$observation_period) == 0) {
+   if(nrow(cdm$observation_period) != 0) {
      cli::cli_inform("The observation period table has been overwritten.")
    }
 
