+++ DESCRIPTION
@@ -46,3 +46,4 @@ Config/testthat/edition: 3
 Encoding: UTF-8
 RoxygenNote: 7.3.2
 BugReports: https://github.com/ohdsi/omock/issues
+LazyData: true
