+++ _pkgdown.yml
@@ -6,7 +6,13 @@ reference:
   - title: "Mock CDM Initialization"
     desc: "This group includes functions that initialize the foundational structures required for the mock CDM environment."
     contents:
-      - matches("mockCdmReference|mockCdmFromTables")
+      - mockCdmReference
+      - mockCdmFromTables
+
+  - title: "Mock Datasets"
+    desc: "This group of functions are utility functions to work with the mock datasets."
+    contents:
+      - mockDatasets
 
   - title: "Mock Table Creation"
     desc: "These functions are focused on adding mock data tables based on the initialized CDM structure."
