### 2025-06-05 e5b4ce3
add functions to download mock datasets
