+++ R/mockDatasets.R
@@ -1,4 +1,75 @@
 
+#' Create a `local` cdm_reference from a dataset.
+#'
+#' @param datasetName Name of the mock dataset. See `availableMockDatasets()`
+#' for possibilities.
+#'
+#' @return A local cdm_reference object.
+#' @export
+#'
+#' @examples
+#' library(omock)
+#'
+#' mockDatasetsFolder(tempdir())
+#' downloadMockDataset(datasetName = "GiBleed")
+#' cdm <- mockCdmFromDataset(datasetName = "GiBleed")
+#' cdm
+#'
+mockCdmFromDataset <- function(datasetName = "GiBleed") {
+  # initial check
+  datasetName <- validateDatasetName(datasetName)
+  cn <- omock::mockDatasets$cdm_name[omock::mockDatasets$dataset_name == datasetName]
+  cv <- omock::mockDatasets$cdm_version[omock::mockDatasets$dataset_name == datasetName]
+
+  # make dataset available
+  datasetPath <- datasetAvailable(datasetName)
+
+  # folder to unzip
+  tmpFolder <- file.path(tempdir(), omopgenerics::uniqueId())
+  if (dir.exists(tmpFolder)) {
+    unlink(x = tmpFolder, recursive = FALSE)
+  }
+  dir.create(tmpFolder)
+
+  # unzip
+  utils::unzip(zipfile = datasetPath, exdir = tmpFolder)
+  tables <- readTables(tmpFolder, cv)
+
+  # delete csv files
+  unlink(x = tmpFolder, recursive = TRUE)
+
+  omopgenerics::cdmFromTables(tables = tables, cdmName = cn, cdmVersion = cv)
+}
+readTables <- function(tmpFolder, cv) {
+  tables <- list.files(tmpFolder, full.names = TRUE, pattern = "\\.csv$")
+  names(tables) <- substr(basename(tables), 1, nchar(basename(tables)) - 4)
+  tables <- as.list(tables)
+  x <- omopgenerics::omopTableFields(cdmVersion = cv)
+  for (nm in names(tables)) {
+    file <- tables[[nm]]
+    cols <- stringr::str_split_1(readLines(con = file, n = 1), ",")
+    colType <- x |>
+      dplyr::filter(
+        .data$cdm_table_name == .env$nm, .data$cdm_field_name %in% .env$cols
+      ) |>
+      dplyr::mutate(cdm_datatype = dplyr::case_when(
+        .data$cdm_datatype == "integer" ~ "i",
+        .data$cdm_datatype == "datetime" ~ "T",
+        .data$cdm_datatype == "date" ~ "D",
+        .data$cdm_datatype == "float" ~ "d",
+        .data$cdm_datatype == "logical" ~ "l",
+        stringr::str_detect(.data$cdm_datatype, "varchar") ~ "c"
+      ))
+    colTypes <- rlang::set_names(as.list(colType$cdm_datatype), colType$cdm_field_name)
+    missingTypes <- cols[!cols %in% names(colTypes)]
+    for (mt in missingTypes) {
+      colTypes[[mt]] <- "?"
+    }
+    tables[[nm]] <- readr::read_csv(file = file, col_types = colTypes)
+  }
+  tables
+}
+
 #' Available mock OMOP CDM Synthetic Datasets
 #'
 #' These are the mock OMOP CDM Synthetic Datasets that are available to download
@@ -18,3 +89,200 @@
 #' mockDatasets
 #'
 "mockDatasets"
+
+#' Download an OMOP Synthetic dataset.
+#'
+#' @param datasetName Name of the mock dataset. See `availableMockDatasets()`
+#' for possibilities.
+#' @param path Path where to download the dataset.
+#' @param overwrite Whether to overwrite the dataset if it is already
+#' downloaded.
+#'
[…]
