+++ man/mockDatasetsStatus.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/mockDatasets.R
+\name{mockDatasetsStatus}
+\alias{mockDatasetsStatus}
+\title{Check the availability of the OMOP CDM datasets.}
+\usage{
+mockDatasetsStatus()
+}
+\value{
+A message with the availability of the datasets.
+}
+\description{
+Check the availability of the OMOP CDM datasets.
+}
+\examples{
+library(omock)
+
+mockDatasetsStatus()
+
+}
