+++ R/mockDatasets.R
@@ -41,31 +41,30 @@ mockCdmFromDataset <- function(datasetName = "GiBleed") {
   omopgenerics::cdmFromTables(tables = tables, cdmName = cn, cdmVersion = cv)
 }
 readTables <- function(tmpFolder, cv) {
-  tables <- list.files(tmpFolder, full.names = TRUE, pattern = "\\.csv$")
-  names(tables) <- substr(basename(tables), 1, nchar(basename(tables)) - 4)
+  tables <- list.files(tmpFolder, full.names = TRUE, pattern = "\\.parquet$", recursive = TRUE)
+  names(tables) <- substr(basename(tables), 1, nchar(basename(tables)) - 8)
   tables <- as.list(tables)
   x <- omopgenerics::omopTableFields(cdmVersion = cv)
   for (nm in names(tables)) {
-    file <- tables[[nm]]
-    cols <- stringr::str_split_1(readLines(con = file, n = 1), ",")
-    colType <- x |>
-      dplyr::filter(
-        .data$cdm_table_name == .env$nm, .data$cdm_field_name %in% .env$cols
-      ) |>
-      dplyr::mutate(cdm_datatype = dplyr::case_when(
-        .data$cdm_datatype == "integer" ~ "i",
-        .data$cdm_datatype == "datetime" ~ "T",
-        .data$cdm_datatype == "date" ~ "D",
-        .data$cdm_datatype == "float" ~ "d",
-        .data$cdm_datatype == "logical" ~ "l",
-        stringr::str_detect(.data$cdm_datatype, "varchar") ~ "c"
-      ))
-    colTypes <- rlang::set_names(as.list(colType$cdm_datatype), colType$cdm_field_name)
-    missingTypes <- cols[!cols %in% names(colTypes)]
-    for (mt in missingTypes) {
-      colTypes[[mt]] <- "?"
+    # read file
+    tables[[nm]] <- arrow::read_parquet(file = tables[[nm]])
+
+    # cast columns
+    xnm <- x |>
+      dplyr::filter(.data$cdm_table_name == .env$nm)
+    for (col in colnames(tables[[nm]])) {
+      type <- xnm$cdm_datatype[xnm$cdm_field_name == col]
+      if (length(type) == 1) {
+        fun <- switch(type,
+                      integer = as.integer,
+                      datetime = as.POSIXct,
+                      date = as.Date,
+                      float = as.numeric,
+                      logical = as.logical,
+                      as.character)
+        tables[[nm]][[col]] <- do.call(fun, list(tables[[nm]][[col]]))
+      }
     }
-    tables[[nm]] <- readr::read_csv(file = file, col_types = colTypes)
   }
   tables
 }
@@ -96,7 +95,7 @@ readTables <- function(tmpFolder, cv) {
 #' for possibilities.
 #' @param path Path where to download the dataset.
 #' @param overwrite Whether to overwrite the dataset if it is already
-#' downloaded.
+#' downloaded. If NULL the used is asked whether to overwrite.
 #'
 #' @return The path to the downloaded dataset.
 #' @export
@@ -112,22 +111,25 @@ readTables <- function(tmpFolder, cv) {
 #'
 downloadMockDataset <- function(datasetName = "GiBleed",
                                 path = mockDatasetsFolder(),
-                                overwrite = FALSE) {
+                                overwrite = NULL) {
   # initial checks
   datasetName <- validateDatasetName(datasetName)
   path <- validatePath(path)
-  omopgenerics::assertLogical(overwrite, length = TRUE)
+  omopgenerics::assertLogical(overwrite, length = 1, null = TRUE)
 
   datasetFile <- file.path(path, paste0(datasetName, ".zip"))
[…]
