+++ DESCRIPTION
@@ -25,18 +25,17 @@ License: Apache License (>= 2)
 URL: https://ohdsi.github.io/omock/, https://github.com/ohdsi/omock
 Depends: 
     R (>= 4.1.0)
-Imports: 
+Imports:
+    arrow,
     cli,
     dplyr,
     lubridate,
     methods,
     omopgenerics (>= 0.3.0),
     purrr,
-    readr,
     rlang,
     snakecase,
-    stats,
-    stringr
+    stats
 Suggests: 
     ggplot2,
     knitr,
