### 2025-06-06 81fa1ab
add action to download all datasets
