+++ .github/workflows/test-all-datasets.yaml
@@ -22,7 +22,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: any::rcmdcheck
+          extra-packages: any::rcmdcheck any::devtools
           needs: check
 
       - name: Test mock datasets
