+++ Datasets/testAllDatasets.R
@@ -1,6 +1,6 @@
 
 devtools::load_all()
-nm <- availableMockDatasets()
+nms <- availableMockDatasets()
 for (nm in nms) {
   time <- format(Sys.time(), "%Y-%m-%d %H:%M:%S")
   cli::cli_inform(c("i" = "{time} Creating Dataset {nm}."))
