### 2025-06-06 103c280
Add drug strength table
