+++ DESCRIPTION
@@ -33,6 +33,7 @@ Imports:
     methods,
     omopgenerics (>= 0.3.0),
     purrr,
+    readr,
     rlang,
     snakecase,
     stats
