### 2025-06-06 2b84aaf
cast mock columns
