### 2025-06-12 ba5b9f2
update
speed up all the mock functions
