+++ NEWS.md
@@ -1,6 +1,21 @@
 # omock (development version)
 
+## What's Changed
+* Create draft-pdf.yml by @catalamarti in https://github.com/OHDSI/omock/pull/140
+* Review paper by @catalamarti in https://github.com/OHDSI/omock/pull/142
+* Update paper.md by @ilovemane in https://github.com/OHDSI/omock/pull/143
+* update by @ilovemane in https://github.com/OHDSI/omock/pull/147
+* fixes doi by @ilovemane in https://github.com/OHDSI/omock/pull/148
+* adding contributing guidlines by @ilovemane in https://github.com/OHDSI/omock/pull/152
+* speed start and end dates by @catalamarti in https://github.com/OHDSI/omock/pull/150
+* Update mockObservationPeriod.R by @ilovemane in https://github.com/OHDSI/omock/pull/153
+* add mockDatasets by @catalamarti in https://github.com/OHDSI/omock/pull/154
+* Update mockCohort.R by @ilovemane in https://github.com/OHDSI/omock/pull/156
+* Test mock datasets by @catalamarti in https://github.com/OHDSI/omock/pull/155
+
 * Add mock datasets by @catalamarti #155 #154
+
+# omock 0.4.0
 
 # omock 0.3.2
 
