### 2025-06-12 0c63ce7
Increment version number to 0.4.0
