+++ README.md
@@ -85,10 +85,10 @@ cdm$person %>%
 #> Rows: 1,000
 #> Columns: 18
 #> $ person_id                   <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13,…
-#> $ gender_concept_id           <int> 8507, 8532, 8532, 8507, 8532, 8507, 8532, …
-#> $ year_of_birth               <int> 1956, 1966, 1983, 1990, 1993, 1983, 1952, …
-#> $ month_of_birth              <int> 5, 8, 12, 10, 11, 7, 2, 3, 6, 11, 11, 2, 7…
-#> $ day_of_birth                <int> 18, 22, 30, 21, 5, 24, 15, 9, 3, 16, 1, 19…
+#> $ gender_concept_id           <int> 8532, 8532, 8532, 8532, 8507, 8532, 8507, …
+#> $ year_of_birth               <int> 1982, 1981, 1962, 1963, 1996, 1972, 1951, …
+#> $ month_of_birth              <int> 5, 4, 5, 1, 10, 4, 11, 1, 7, 12, 10, 5, 1,…
+#> $ day_of_birth                <int> 3, 8, 18, 25, 17, 7, 12, 5, 7, 7, 1, 28, 8…
 #> $ race_concept_id             <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
 #> $ ethnicity_concept_id        <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
 #> $ birth_datetime              <dttm> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
@@ -116,7 +116,7 @@ cdm$observation_period %>%
 #> Columns: 5
 #> $ observation_period_id         <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 1…
 #> $ person_id                     <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 1…
-#> $ observation_period_start_date <date> 1993-06-25, 2003-07-30, 2000-05-24, 200…
-#> $ observation_period_end_date   <date> 2002-11-11, 2015-10-12, 2014-07-08, 201…
+#> $ observation_period_start_date <date> 1985-10-17, 2004-06-19, 1988-06-21, 201…
+#> $ observation_period_end_date   <date> 2002-10-01, 2006-01-25, 2017-08-08, 201…
 #> $ period_type_concept_id        <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
 ```
