+++ NEWS.md
@@ -1,3 +1,5 @@
+# omock (development version)
+
 # omock 0.4.0
 
 ## What's Changed
