### 2025-06-16 e7d8807
Create copyCdmToDuckdb
