+++ R/copyCdmToDuckdb.R
@@ -1,0 +1,183 @@
+
+#' Copy a local cdm to a duckdb connection
+#'
+#' `r lifecycle::badge('experimental')`
+#'
+#' This function is a temporary placeholder and will be moved in a more suitable
+#' package in the future.
+#'
+#' @param cdm A cdm_reference object. If not local it fill be collected.
+#' @param db A duckdb connection or a path to create a duckdb database. If NULL
+#' the database will be created in a temporary file.
+#' @param cdmSchema String, character vector with the schema to write the cdm
+#' tables.
+#' @param writeSchema String, character vector with the schema to write the
+#' cohort tables.
+#' @param writePrefix String, character vector with the writing prefix for
+#' cohort and other tables.
+#' @param achillesSchema String, character vector with the schema to write the
+#' achilles tables.
+#' @param achillesPrefix String, character vector with the writing prefix for
+#' achilles tables.
+#'
+#' @return A cdm_reference in a duckdb connection.
+#' @export
+#'
+#' @examples
+#' \donttest{
+#' library(omock)
+#'
+#' cdm <- mockCdmFromDataset(datasetName = "GiBleed")
+#' duckcdm <- copyCdmToDuckdb(cdm = cdm)
+#' duckcdm
+#'
+#' duckcdm2 <- copyCdmToDuckdb(cdm = cdm, db = tempdir())
+#' duckcdm2
+#'
+#' }
+copyCdmToDuckdb <- function(cdm,
+                            db = NULL,
+                            cdmSchema = "main",
+                            writeSchema = "results",
+                            writePrefix = "",
+                            achillesSchema = "results") {
+  rlang::check_required("duckdb")
+  rlang::check_required("CDMConnector")
+
+  # initial validation
+  cdm <- omopgenerics::validateCdmArgument(cdm = cdm)
+  con <- validateDb(db = db, cdmName = omopgenerics::cdmName(x = cdm))
+  cdmSchema <- validateSchema(schema = cdmSchema, con = con)
+  writeSchema <- validateSchema(schema = writeSchema, con = con)
+  omopgenerics::assertCharacter(writePrefix, length = 1)
+  achillesSchema <- validateSchema(schema = achillesSchema, con = con)
+
+  tables <- cdmTablesClasses(cdm = cdm)
+
+  # write cdm tables
+  for (nm in tables$omop_tables) {
+    x <- dplyr::collect(cdm[[nm]])
+    writeTable(x = x, con = con, schema = cdmSchema, prefix = "", name = nm)
+  }
+
+  # write cohort tables
+  for (nm in tables$cohort_tables) {
+    x <- dplyr::collect(cdm[[nm]])
+    list(x, attr(x, "cohort_set"), attr(x, "cohort_attrition"), attr(x, "cohort_codelist")) |>
+      rlang::set_names(nm = paste0(nm, c("", "_set", "_attrition", "_codelist"))) |>
+      purrr::imap(\(x, nm) {
+        writeTable(x = x, con = con, schema = writeSchema, prefix = writePrefix, name = nm)
+      })
+  }
+
+  # write achilles tables
+  for (nm in tables$achilles_tables) {
+    x <- dplyr::collect(cdm[[nm]])
+    writeTable(x = x, con = con, schema = achillesSchema, prefix = "", name = nm)
+  }
+  if (length(tables$achilles_tables) == 0) {
+    achillesSchema <- NULL
+  }
+
+  # write other tables
+  for (nm in tables$other_tables) {
+    x <- dplyr::collect(cdm[[nm]])
[…]
