### 2025-06-16 49b28c4
Edit NEWS.md
