+++ man/mockDrugExposure.Rd
@@ -7,7 +7,7 @@
 mockDrugExposure(cdm, recordPerson = 1, seed = NULL)
 }
 \arguments{
-\item{cdm}{A `cdm_reference` object that must already include 'person' and 'observation_period' tables.
+\item{cdm}{A \code{cdm_reference} object that must already include 'person' and 'observation_period' tables.
 This object serves as the base CDM structure where the drug exposure data will be added.
 The 'person' and 'observation_period' tables must be populated as they are necessary for generating accurate drug exposure records.}
 
@@ -20,9 +20,9 @@ If provided, this seed enables the function to produce consistent results each t
 If 'NULL', the seed is not set, which can lead to different outputs on each run.}
 }
 \value{
-Returns the modified `cdm` object with the new 'drug_exposure' table added. This table includes the simulated
-        drug exposure data for each person, ensuring that each record is correctly linked to individuals in the 'person' table
-        and falls within valid observation periods.
+Returns the modified \code{cdm} object with the new 'drug_exposure' table added. This table includes the simulated
+drug exposure data for each person, ensuring that each record is correctly linked to individuals in the 'person' table
+and falls within valid observation periods.
 }
 \description{
 This function simulates drug exposure records for individuals within a specified cohort. It creates a realistic dataset by generating drug exposure records based on the specified number of records per person. Each drug exposure record is correctly associated with an individual within valid observation periods, ensuring the integrity of the data.
