+++ man/mockObservationPeriod.Rd
@@ -7,7 +7,7 @@
 mockObservationPeriod(cdm, seed = NULL)
 }
 \arguments{
-\item{cdm}{A `cdm_reference` object that must include a 'person' table with valid dates of birth.
+\item{cdm}{A \code{cdm_reference} object that must include a 'person' table with valid dates of birth.
 This object serves as the base CDM structure where the observation period data will be added.
 The function checks to ensure that the 'person' table is populated and uses the date of birth to generate observation periods.}
 
@@ -16,8 +16,8 @@ If provided, this seed allows the function to produce consistent results each ti
 If 'NULL', the seed is not set, which can lead to different outputs on each run.}
 }
 \value{
-Returns the modified `cdm` object with the new 'observation_period' table added. This table includes the simulated
-        observation periods for each person, ensuring that each record spans a realistic timeframe based on the person's date of birth.
+Returns the modified \code{cdm} object with the new 'observation_period' table added. This table includes the simulated
+observation periods for each person, ensuring that each record spans a realistic timeframe based on the person's date of birth.
 }
 \description{
 This function simulates observation periods for individuals based on their date of birth recorded in the 'person' table of the CDM object. It assigns random start and end dates for each observation period within a realistic timeframe up to a specified or default maximum date.
