+++ man/mockMeasurement.Rd
@@ -7,7 +7,7 @@
 mockMeasurement(cdm, recordPerson = 1, seed = NULL)
 }
 \arguments{
-\item{cdm}{A `cdm_reference` object that must already include 'person' and 'observation_period' tables.
+\item{cdm}{A \code{cdm_reference} object that must already include 'person' and 'observation_period' tables.
 This object serves as the base CDM structure where the measurement data will be added.
 The 'person' and 'observation_period' tables must be populated as they are necessary for generating accurate measurement records.}
 
@@ -20,9 +20,9 @@ If provided, this seed enables the function to produce consistent results each t
 If 'NULL', the seed is not set, which can lead to different outputs on each run.}
 }
 \value{
-Returns the modified `cdm` object with the new 'measurement' table added. This table includes the simulated
-        measurement data for each person, ensuring that each record is correctly linked to individuals in the 'person' table
-        and falls within valid observation periods.
+Returns the modified \code{cdm} object with the new 'measurement' table added. This table includes the simulated
+measurement data for each person, ensuring that each record is correctly linked to individuals in the 'person' table
+and falls within valid observation periods.
 }
 \description{
 This function simulates measurement records for individuals within a specified cohort. It creates a realistic dataset by generating measurement records based on the specified number of records per person. Each measurement record is correctly associated with an individual within valid observation periods, ensuring the integrity of the data.
