+++ man/copyCdmToDuckdb.Rd
@@ -1,0 +1,59 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/copyCdmToDuckdb.R
+\name{copyCdmToDuckdb}
+\alias{copyCdmToDuckdb}
+\title{Copy a local cdm to a duckdb connection}
+\usage{
+copyCdmToDuckdb(
+  cdm,
+  db = NULL,
+  cdmSchema = "main",
+  writeSchema = "results",
+  writePrefix = "",
+  achillesSchema = "results"
+)
+}
+\arguments{
+\item{cdm}{A cdm_reference object. If not local it fill be collected.}
+
+\item{db}{A duckdb connection or a path to create a duckdb database. If NULL
+the database will be created in a temporary file.}
+
+\item{cdmSchema}{String, character vector with the schema to write the cdm
+tables.}
+
+\item{writeSchema}{String, character vector with the schema to write the
+cohort tables.}
+
+\item{writePrefix}{String, character vector with the writing prefix for
+cohort and other tables.}
+
+\item{achillesSchema}{String, character vector with the schema to write the
+achilles tables.}
+
+\item{achillesPrefix}{String, character vector with the writing prefix for
+achilles tables.}
+}
+\value{
+A cdm_reference in a duckdb connection.
+}
+\description{
+\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
+}
+\details{
+This function is a temporary placeholder and will be moved in a more suitable
+package in the future.
+}
+\examples{
+\donttest{
+library(omock)
+
+cdm <- mockCdmFromDataset(datasetName = "GiBleed")
+duckcdm <- copyCdmToDuckdb(cdm = cdm)
+duckcdm
+
+duckcdm2 <- copyCdmToDuckdb(cdm = cdm, db = tempdir())
+duckcdm2
+
+}
+}
