+++ NAMESPACE
@@ -7,6 +7,7 @@ export(cdmName)
 export(cdmSource)
 export(cdmVersion)
 export(cohortCount)
+export(copyCdmToDuckdb)
 export(downloadMockDataset)
 export(dropTable)
 export(emptyCdmReference)
