+++ R/copyCdmToDuckdb.R
@@ -17,8 +17,6 @@
 #' cohort and other tables.
 #' @param achillesSchema String, character vector with the schema to write the
 #' achilles tables.
-#' @param achillesPrefix String, character vector with the writing prefix for
-#' achilles tables.
 #'
 #' @return A cdm_reference in a duckdb connection.
 #' @export
