### 2025-06-23 3ac790e
update documentation for vocabset arguement
