+++ paper/paper.md
@@ -13,11 +13,11 @@ authors:
   - name: Núria Mercadé-Besora
     orcid: 0009-0006-7948-3747
     affiliation: 1
-  - name: Xihang Chen
-    orcid: 0009-0001-8112-8959
-    affiliation: 1
   - name: Marta Alcalde-Herraiz
     orcid: 0009-0002-4405-1814
+    affiliation: 1
+  - name: Xihang Chen
+    orcid: 0009-0001-8112-8959
     affiliation: 1
   - name: Yuchen Guo
     orcid: 0000-0002-0847-4855
@@ -49,7 +49,7 @@ Affiliation:\
 
 # Summary
 
-`omock` is an R package [@inproceedings] that allows users to create mock patient level data in the formatted in the Observational Medical Outcomes Partnership (OMOP) Common Data Model (CDM) [@omop]. This package provides a flexible and efficient way to create synthetic datasets in OMOP CDM format, facilitating the testing and validation of packages and analytic codes.
+`omock` is an R package that allows users to create mock patient level data in the formatted in the Observational Medical Outcomes Partnership (OMOP) Common Data Model (CDM) [@omop]. This package provides a flexible and efficient way to create synthetic datasets in OMOP CDM format, facilitating the testing and validation of packages and analytic codes.
 
 # Statement of need
 
@@ -87,6 +87,8 @@ An empty mock CDM can be created using the `mockCdmReference` function, which in
 
 To expand the mock CDM with additional clinical tables, we can pipe the corresponding functions onto the previously created CDM object. For example, to add a condition occurrence and a drug exposure table to the OMOP CDM we can use `mockConditionOccurrence` and `mockDrugExposure` functions, respectively. There is a function for most of the commonly used clinical tables in OMOP CDM.
 
+This modular design, where each table is added via a separate function, provides flexibility in testing and development workflows. Since users can simulate the specific OMOP CDM tables as needed, depending on the scenario being tested, such as testing functionality that only requires drug exposures or condition occurrences. Generating only the necessary tables can be less computationally intensive, which is especially useful when simulating CDM with a large number of patients or when running automated test suites repeatedly. Additionally, table-level control also makes it easier to simulate edge cases and specify population settings, such as the number of patients, gender and number of records within each tables.
+
 Below is an example code snippet demonstrating how to generate a mock CDM with 1,000 patients, valid observation periods, and additional drug exposure and condition occurrence tables:
 
 ```         
@@ -120,8 +122,6 @@ print(cdm)
 Similarly, the `mockCohort` function can add a mock cohort table to the mock CDM object. The function contains arguments for the user to customise the mock table, such as its name and size. See the example below.
 
 ```         
-library(omock)
-
 cdm <- cdm |>
   mockCohort(
     name = "omock_example",
@@ -149,7 +149,13 @@ print(cdm)
 
 ## Building mock CDM object with bespoke CDM table
 
-To create a mock CDM object from a bespoke OMOP table, we can use the `mockCdmFromTables` function. `mockCdmFromTables` takes the OMOP table in `tibble` format and create a CDM object with valid person and observation period information for those tables. Below is an example code snippet on using `mockCdmFromTables` to create a mock CDM object from the `condition_occurrence` table.
[…]
