+++ paper/paper.bib
@@ -152,3 +152,12 @@ doi = {10.1109/ICSE43902.2021.00136}
   year={2019},
   publisher={OHDSI}
 }
+
+@Manual{eunomia,
+  title = {Eunomia: Standard Dataset Manager for Observational Medical Outcomes Partnership Common Data Model Sample Datasets},
+  author = {Frank DeFalco and Martijn Schuemie and Anthony Sena and Natthawut Adulyanukosol and Star Liu and Adam Black},
+  year = {2024},
+  note = {R package version 2.0.0},
+  url = {https://github.com/OHDSI/Eunomia/},
+  doi = {10.32614/CRAN.package.Eunomia}
+}
