### 2025-07-09 f50daaa
Update paper.md
