### 2025-07-09 #8178 @ilovemane
Dear @louisahsmith,

Thank you very much for your time to review my package paper, the comments are really useful. I have tried my best to address them as much as I can below. Any more concern please let me know.

best wishes,

Mike

- I found the documentation lacking in some places, e.g., options for vocabularySet argument in mockCdmReference() (what does it mean to define the vocabulary tables to be included (line 73) -- there is a different description of the argument in omock::mockVocabularyTables() and it looks in the source code like it can only be "mock" or "eunomia," though I can't find this in the documentation. What are the implications of choosing one over the other?

"I’ve expanded the description for the vocabularySet argument and submitted a pull request here where you can review the changes. https://github.com/OHDSI/omock/pull/162

In simple terms, the vocabularySet argument in mockCdmReference() defines which set of vocabulary tables the mock CDM will use. Currently, it supports two options:

"eunomia": uses vocabulary tables derived from the Eunomia dataset
"mock": uses synthetic, non-realistic vocabularies intended for testing scenarios that don't require valid concept. The "mock" option is especially useful for fast unit tests or package checks where vocabulary isn’t essential."

- There is a missing vignette on the website (https://ohdsi.github.io/omock/articles/a03_Creating_a_synthetic_vocabulary.html) that would be helpful to point readers toward when the paper is published.

Thank you for spotting this, I have finished the vignette, this should also make the vocabulary table creation more clear to users. You can review the pull request here where you can see the vignette. https://github.com/OHDSI/omock/pull/160

- The Vidoni 2021 paper is cited in the statement of need, where it makes sense, but I'm not sure it is what is meant to be cited in the first sentence of the summary?

Thanks for spotting this, I have got rid if this now. You can view it in this pull request line 52 https://github.com/OHDSI/omock/pull/167/files#diff-e504eb580b095a7e65428b098183a581e475f0fb316db95287eacd7d4f344424

- Line 111 -- it took me a little bit to figure out why this code wrote over the previous cdm object -- of course it's because it's just adding a cohort and maintaining everything about the cdm (just like the tables were added in the first piped section). However, I was reading the code more closely than the text at first, and the name of the function led me to believe that this would be some kind of cohort standalone object. It would likely be too major a change to change all the function names to, e.g., **add**MockCohort() once you are building off an initial mockCdmReference(), but perhaps this could be made clearer in other ways?
Also, perhaps don't call library(omock) again on line 109 to make it clear it is a continuation of the earlier code?
Then, in the second example (using mockCdmFromTables()), I'd suggest naming it something other than cdm to distinguish this example, which starts from scratch -- it could be confusing about whether you are building on to what was previously created

Thank you for this, we did consider changing the naming of the functions that add extra table to the cdm. However we cannot come up with a naming that are not too long, appropriate and logical. We cannot use **add** because omock was part of a group of packages the research group I worked in https://oxford-pharmacoepi.github.io/Oxinfer/packages/summary.html. The *add* were use in the PatientProfiles package for adding columns to a dataset. So we dont want to confused users from that. At the moment in omock, if the function contain "cdm" it will create a new cdm, if they dont contain "cdm", the function would require a cdm object as input and add an extra table in the cdm object provided.

I have amended line 109 and changed the naming in the example in the paper as you suggested. You can view in this pull request https://github.com/OHDSI/omock/pull/167/files in line 124 and 174

- Along those lines, what are the benefits of using separate functions to continue to add tables to the cdm (e.g., a drug exposure table) as you go, rather than initializing them all in the initial function (as is the case for the specified tables in the mockCdmFromTables() function)? This became clearer to me after playing around with the package, but I think answering it would be helpful in the paper.

I have added a paragraph to explain the motivation behind this in the paper. https://github.com/OHDSI/omock/pull/167/files line 90

"This modular design, where each table is added via a separate function, provides flexibility in testing and development workflows. Since users can simulate the specific OMOP CDM tables as needed, depending on the scenario being tested, such as testing functionality that only requires drug exposures or condition occurrences. Generating only the necessary tables can be less computationally intensive, which is especially useful when simulating CDM with a large number of patients or when running automated test suites repeatedly. Additionally, table-level control also makes it easier to simulate edge cases and specify population settings, such as the number of patients, gender and number of records within each tables."

- I would not have guessed from the code alone that the second cdm (initialized with mockCdmFromTables()) would have a person table and an observation period table, as those were not specified as arguments nor added on with the mockPerson(), etc. functions as in the first example. I only understood after looking into the documentation a little bit more---it was not clear to me that the first argument (not provided in the example) was actually a cdm reference with a default value of mockCdmReference(). I looked into this very briefly and could not find out how, for example, I would then specify the other arguments to describe the people in the person table (e.g., proportion female). Or, for example, what if you specify the condition table but also want some people who don't have any conditions (but don't want to fully specify the person table)? I could not edit the person table, e.g., by running mockPerson() on the resulting cdm, and when I tried to add a manual condition occurrence table to an existing cdm object with a person table, it overwrote that. Maybe this is not a major concern, but this is hard to know without more concrete examples of how the package can be used.

Thank you for pointing this out. mockCdmReference() was meant to use in scenario where the users have to create a unit test require a bespoke clinical/cohort tables in a cdm object, without observation_period and person requirement. mockCdmReference() will take those bespoke tables as input and create a cdm object with valid person and observation period from the person_id from the input and dates of the input tables. I have added a section in the paper "Testing code with omock" and provided examples of a scenario when mockCdmReference() will be useful. https://github.com/OHDSI/omock/pull/167/files line 197

- There is a big emphasis on how the package can help with testing. It would be helpful to describe more about the arguments that can be used to tailor your cdm toward your needs, and maybe an example of how it fits into a testing pipeline? Right now, I can see how it would be useful for tests that require certain tables to exist, etc. But it would be helpful to describe a more complex situation where you need a table that meets certain conditions. When would changing the gender proportions or the dates matter, i.e., why have you chosen to include this functionality?

I have added a section in the paper "Testing code with omock" and provided examples of a scenario when mockCdmReference() will be useful and a scenario when mockCdmReference() should be use. https://github.com/OHDSI/omock/pull/167/files line 197

- Along those lines, if you provide your own tables, does omock help make sure they are internally consistent (e.g., dates falling within the observation period)? For example, scrolling through the source code, I see a function called correctDateDeath(). While there's no need to mention or document this non-user-facing function in particular, it would be helpful to know more about these types of checks/corrections, both to guide potential users as well as to highlight your innovations with the package.
Stylistically/for consistency, could change the assignment = to the assignment <- on line 140

Thank you for the feedback. The omock package itself doesn’t include internal functions to validate whether the generated tables strictly follow OMOP CDM conventions. Instead, it leverages the omopgenerics package, developed by our research group, which contains all the necessary validation functions to ensure compliance with OMOP CDM structure and rules.

This is mentioned in line 70 of the paper, where we describe omopgenerics as a core dependency. All tables generated by omock are constructed to be valid according to these rules, using the structure and helper functions defined in omopgenerics.

- In the source code of mockCdmFromTables(), there is a hardcoded date ("01-01-2024") that is already out of date and I can imagine will get more and more so. Changing the date in the last test of this function from "2020-01-01" to "2023-12-01" will provoke this error.

Thank you for pointing this out. I’ve now updated the function to replace the hardcoded date ("01-01-2024") with a user-defined argument, allowing users to specify the maximum allowable observation_period_end_date. This change is intended to ensure that the generated mock data is more realistic and does not contain dates set in the future.

- Looks like good test coverage, and all the tests passed, but this function was also the source of 4 warnings about "no non-missing arguments to minmax"

This indeed is an error. I wanted mockCdmFromTables to be able to take in observation_period table as input and return a cdm with valid person table. I have amended the function and tests. The changes you can see in this pull request. https://github.com/OHDSI/omock/pull/168/files

- I imagine some people use Eunomia to test packages that use OMOP data. It would be helpful compare and contrast with this package.

I have added a new section called comparison with existing tools in the paper, to explain this. https://github.com/OHDSI/omock/pull/167/files line 275

"A commonly used resource for testing OMOP-based tools is the `Eunomia` package [@eunomia], which provides a fixed, small synthetic CDM dataset. `Eunomia` is a valuable tool for exploring the OMOP CDM data structure and for demonstration purposes. It is less suited for testing code that requires custom scenarios, such as examples 1 and 2 shown above. Since the data within the `Eunomia` package is static and cannot be easily modified, this made it challenging to generate custom mock data for unit testing. Whereas `omock` is designed to give users the flexibility for generating mock OMOP CDM data tailored to their testing needs."
