### 2025-07-09 #8178 @ilovemane
Dear @chitrams,

Thank you very much for your time to review my package paper, the comments are really useful. I have tried my best to address them as much as I can below. If you have any further concerns, please don't hesitate to let me know.

best wishes,

Mike

📝 editing: In the first line in Summary: "omock is an R package that allows users to create mock patient level data in the formatted in the Observational Medical Outcomes Partnership (OMOP) ..."

This have been corrected. line 52 in this pull request. https://github.com/OHDSI/omock/pull/167

Consider snake case for function names instead of camel case, as per Tidyverse [style guide](https://style.tidyverse.org/syntax.html), but this is very much a matter of preference!

Thank you for the suggestion. However, I’m unable to change the naming convention to snake case, as this package is part of a broader set of packages developed by our research group. As a group, we've agreed to use camel case for function names to maintain consistency across all packages in the ecosystem (see: [Oxinfer package suite](https://oxford-pharmacoepi.github.io/Oxinfer/packages/summary.html)). I hope you understand the need to follow this shared standard.
