### 2025-07-10 4ae0427
add duckdb as source
