+++ NAMESPACE
@@ -7,7 +7,6 @@ export(cdmName)
 export(cdmSource)
 export(cdmVersion)
 export(cohortCount)
-export(copyCdmToDuckdb)
 export(downloadMockDataset)
 export(dropTable)
 export(emptyCdmReference)
