+++ NAMESPACE
@@ -28,6 +28,7 @@ export(mockObservationPeriod)
 export(mockPerson)
 export(mockProcedureOccurrence)
 export(mockVisitOccurrence)
+export(mockVocabularySet)
 export(mockVocabularyTables)
 export(settings)
 importFrom(omopgenerics,attrition)
