+++ tests/testthat/test-mockVocabularySet.R
@@ -1,10 +1,10 @@
 test_that("mockVocabularySet", {
 
   dbName <- "GiBleed"
-  expect_no_error(cdm <- mockVocabularySet(datasetName = dbName))
+  expect_no_error(cdm <- mockVocabularySet(vocabularySet = dbName))
   expect_no_error(omopgenerics::validateCdmArgument(cdm))
 
-  expect_no_error(cdm <- mockVocabularySet(datasetName = dbName))
+  expect_no_error(cdm <- mockVocabularySet(vocabularySet = dbName))
 
   expect_equal(cdm$person |> nrow(),0)
 
