+++ tests/testthat/test-mockVocabularySet.R
@@ -1,5 +1,8 @@
 test_that("mockVocabularySet", {
 
+  Sys.setenv("MOCK_DATASETS_FOLDER" = "")
+  myFolder <- file.path(tempdir(), "DATASETS")
+
   dbName <- "GiBleed"
   expect_no_error(cdm <- mockVocabularySet(vocabularySet = dbName))
   expect_no_error(omopgenerics::validateCdmArgument(cdm))
@@ -9,5 +12,7 @@ test_that("mockVocabularySet", {
   expect_equal(cdm$person |> nrow(),0)
 
   expect_equal(cdm$observation_period |> nrow(),0)
+
+  unlink(myFolder, recursive = TRUE)
 
 })
