### 2025-08-08 #8178 @jromanowska
Good to hear from you, @chitrams!
I understand that you have had a lot of work. Thank you for clarifying and we'll wait until the end of August. You've done already a lot on the review, so it'd be wonderful if you can wrap it up. 🙌
In the meantime, have a great weekend everyone!
