### 2025-08-14 #8178 @ilovemane
@jromanowska I saw @louisahsmith re-generated the manuscript, however it seem is still the old one without my edits from the comments. I wonder should I merge the pull-request for the changes I made to the manuscript so the latest one get generated? Or are there command we can use to generate the mnauscript from a different branch?
