### 2025-08-21 #8178 @louisahsmith
Thanks for generating the pdf @ilovemane -- I realized that was the old version so was looking through the version on the other branch but this way was much easier for me to read and follow the changes. All looks good! I appreciate the additional code and text in reponse to my feedback and have checked off all the items in my checklist now.
