+++ README.md
@@ -85,10 +85,10 @@ cdm$person %>%
 #> Rows: 1,000
 #> Columns: 18
 #> $ person_id                   <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13,…
-#> $ gender_concept_id           <int> 8532, 8532, 8532, 8532, 8507, 8532, 8507, …
-#> $ year_of_birth               <int> 1982, 1981, 1962, 1963, 1996, 1972, 1951, …
-#> $ month_of_birth              <int> 5, 4, 5, 1, 10, 4, 11, 1, 7, 12, 10, 5, 1,…
-#> $ day_of_birth                <int> 3, 8, 18, 25, 17, 7, 12, 5, 7, 7, 1, 28, 8…
+#> $ gender_concept_id           <int> 8532, 8532, 8507, 8532, 8507, 8532, 8532, …
+#> $ year_of_birth               <int> 1988, 1994, 1996, 2000, 1973, 1970, 1986, …
+#> $ month_of_birth              <int> 3, 3, 5, 3, 11, 1, 4, 7, 2, 9, 4, 5, 2, 7,…
+#> $ day_of_birth                <int> 28, 23, 20, 14, 8, 26, 1, 21, 23, 2, 7, 31…
 #> $ race_concept_id             <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
 #> $ ethnicity_concept_id        <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
 #> $ birth_datetime              <dttm> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
@@ -116,7 +116,7 @@ cdm$observation_period %>%
 #> Columns: 5
 #> $ observation_period_id         <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 1…
 #> $ person_id                     <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 1…
-#> $ observation_period_start_date <date> 1985-10-17, 2004-06-19, 1988-06-21, 201…
-#> $ observation_period_end_date   <date> 2002-10-01, 2006-01-25, 2017-08-08, 201…
+#> $ observation_period_start_date <date> 2000-12-12, 2003-12-17, 2004-10-05, 200…
+#> $ observation_period_end_date   <date> 2010-11-16, 2019-08-27, 2019-05-12, 201…
 #> $ period_type_concept_id        <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
 ```
