+++ revdep/cran.md
@@ -1,7 +1,14 @@
 ## revdepcheck results
 
-We checked 6 reverse dependencies, comparing R CMD check results across CRAN and dev versions of this package.
+We checked 9 reverse dependencies, comparing R CMD check results across CRAN and dev versions of this package.
 
  * We saw 0 new problems
- * We failed to check 0 packages
+ * We failed to check 3 packages
 
+Issues with CRAN packages are summarised below.
+
+### Failed to check
+
+* CodelistGenerator (NA)
+* CohortConstructor (NA)
+* PhenotypeR        (NA)
