+++ revdep/failures.md
@@ -1,1 +1,71 @@
-*Wow, no problems at all. :)*
+# CodelistGenerator
+
+<details>
+
+* Version: 3.5.0
+* GitHub: NA
+* Source code: https://github.com/cran/CodelistGenerator
+* Date/Publication: 2025-04-10 19:20:39 UTC
+* Number of recursive dependencies: 124
+
+Run `revdepcheck::revdep_details(, "CodelistGenerator")` for more info
+
+</details>
+
+## In both
+
+*   R CMD check timed out
+    
+
+# CohortConstructor
+
+<details>
+
+* Version: 0.5.0
+* GitHub: https://github.com/OHDSI/CohortConstructor
+* Source code: https://github.com/cran/CohortConstructor
+* Date/Publication: 2025-07-30 18:30:02 UTC
+* Number of recursive dependencies: 164
+
+Run `revdepcheck::revdep_details(, "CohortConstructor")` for more info
+
+</details>
+
+## In both
+
+*   R CMD check timed out
+    
+
+*   checking installed package size ... NOTE
+    ```
+      installed size is  6.1Mb
+      sub-directories of 1Mb or more:
+        doc   5.1Mb
+    ```
+
+# PhenotypeR
+
+<details>
+
+* Version: 0.2.0
+* GitHub: NA
+* Source code: https://github.com/cran/PhenotypeR
+* Date/Publication: 2025-07-31 10:00:37 UTC
+* Number of recursive dependencies: 197
+
+Run `revdepcheck::revdep_details(, "PhenotypeR")` for more info
+
+</details>
+
+## In both
+
+*   R CMD check timed out
+    
+
+*   checking installed package size ... NOTE
+    ```
+      installed size is  5.1Mb
+      sub-directories of 1Mb or more:
+        doc   4.4Mb
+    ```
+
