### 2025-09-01 20c2cf8
Increment version number to 0.5.0
