+++ NEWS.md
@@ -1,4 +1,4 @@
-# omock (development version)
+# omock 0.5.0
 
 * add option source to `mockCdmFromDataset` #158
 
