### 2025-09-01 627872a
Increment version number to 0.5.0.9000
