+++ NEWS.md
@@ -1,3 +1,5 @@
+# omock (development version)
+
 # omock 0.5.0
 
 * add option source to `mockCdmFromDataset` #158
