### 2025-09-01 ce16c35
Update paper.md
