### 2025-09-02 #8178 @jromanowska
@ilovemane, please remove the "Authors" and "Affiliations" lines from paper (lines 9-14).

Do you have a reference (or webpage) to OMOP CDM (line 31)? I think that might be useful.

Then, regenerate the pdf (you don't need to create any new releases).
