+++ paper/paper.bib
@@ -161,3 +161,8 @@ doi = {10.1109/ICSE43902.2021.00136}
   url = {https://github.com/OHDSI/Eunomia/},
   doi = {10.32614/CRAN.package.Eunomia}
 }
+
+@misc{omopweb,
+title =  OMOP Common Data Model,
+url={https://ohdsi.github.io/CommonDataModel/},
+journal={OMOP Common Data Model}}
