### 2025-09-02 f368780
Update paper.bib
