+++ paper/paper.bib
@@ -163,6 +163,7 @@ doi = {10.1109/ICSE43902.2021.00136}
 }
 
 @misc{omopweb,
-title =  OMOP Common Data Model,
+title = {OMOP Common Data Model},
 url={https://ohdsi.github.io/CommonDataModel/},
-journal={OMOP Common Data Model}}
+journal={OMOP Common Data Model}
+}
