### 2025-09-02 1cb95da
Update paper.md
